requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
opencv-python==4.7.0.72
pdf417decoder==1.0.8
Pillow==9.5.0
Werkzeug==2.3.6
